+++ src/utils.ts
@@ -189,17 +189,27 @@ export function areMapsEqual(
   let isValueEqual = a.size === b.size;
 
   if (isValueEqual && a.size) {
+    const matchedIndices: Record<number, true> = {};
+
     a.forEach((aValue, aKey) => {
       if (isValueEqual) {
-        isValueEqual = false;
-        let iterationIndex = 0;
+        let hasMatch = false;
+        let matchIndex = 0;
 
         b.forEach((bValue, bKey) => {
-          if (!isValueEqual && isEqual(aKey, bKey, iterationIndex, a, b, meta)) {
-            isValueEqual = isEqual(aValue, bValue, aKey, a, b, meta);
+          if (!hasMatch && !matchedIndices[matchIndex]) {
+            hasMatch =
+            isEqual(aKey, bKey, matchIndex, a, b, meta) && isEqual(aValue, bValue, aKey, a, b, meta);
+
+            if (hasMatch) {
+              matchedIndices[matchIndex] = true;
+            }
           }
-          iterationIndex++;
+
+          matchIndex++;
         });
+
+        isValueEqual = hasMatch;
       }
     });
   }
@@ -306,17 +316,26 @@ export function areSetsEqual(
   let isValueEqual = a.size === b.size;
 
   if (isValueEqual && a.size) {
+    const matchedIndices: Record<number, true> = {};
+
     a.forEach((aValue) => {
       if (isValueEqual) {
-        isValueEqual = false;
-        let iterationIndex = 0;
+        let hasMatch = false;
+        let matchIndex = 0;
 
         b.forEach((bValue) => {
-          if (!isValueEqual) {
-            isValueEqual = isEqual(aValue, bValue, iterationIndex, a, b, meta);
+          if (!hasMatch && !matchedIndices[matchIndex]) {
+            hasMatch = isEqual(aValue, bValue, matchIndex, a, b, meta);
+
+            if (hasMatch) {
+              matchedIndices[matchIndex] = true;
+            }
           }
-          iterationIndex++;
+
+          matchIndex++;
         });
+
+        isValueEqual = hasMatch;
       }
     });
   }